requirements.txt
tqdm
datasets
python-dotenv
openai
backoff
# only needed for upload_to_hf
huggingface_hub